add class CountConverter for count existing user's clocks fixed problem with displaying user name add basic scroll Temporary(?) solution for dividing methods to View and Model
