+++ packages.config
@@ -6,6 +6,7 @@
   <package id="Microsoft.CodeCoverage" version="15.9.0" targetFramework="net461" />
   <package id="Microsoft.NET.Test.Sdk" version="15.9.0" targetFramework="net461" />
   <package id="System.ValueTuple" version="4.5.0" targetFramework="net461" />
+  <package id="System.Windows.Interactivity.WPF" version="2.0.20525" targetFramework="net461" />
   <package id="TestStack.White" version="0.13.3" targetFramework="net461" />
   <package id="xunit" version="2.4.0" targetFramework="net461" />
   <package id="xunit.abstractions" version="2.0.3" targetFramework="net461" />
