+++ src/Views/MainView.xaml
@@ -3,12 +3,16 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:wi="clr-namespace:System.Windows.Interactivity;assembly=System.Windows.Interactivity"
+        xmlns:i="http://schemas.microsoft.com/expression/2010/interactions"
+        xmlns:misc="clr-namespace:AlarmClock.Misc"
         mc:Ignorable="d"
-        Height="Auto" Width="Auto" MinHeight="600" MinWidth="800">
+        Height="Auto" Width="Auto" MinHeight="200" MinWidth="800">
     <UserControl.Resources>
+        <misc:CountConverter x:Key="CountConverter"/>
         <BooleanToVisibilityConverter x:Key="BoolToVisibility" />
         <SolidColorBrush x:Key="BackgroundBrush" Color="White"/>
-        <Storyboard x:Key="Blinking" RepeatBehavior="0:0:10">
+        <Storyboard x:Key="Blinking" RepeatBehavior="0:2:00">
             <DoubleAnimation Storyboard.TargetProperty="Opacity" AutoReverse="True"
                 To="0.1"
                 Duration="0:0:2.0"
@@ -49,7 +53,14 @@
             <StackPanel Orientation="Vertical">
                 <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="0,2">
                     <StackPanel Orientation="Horizontal" IsEnabled="{Binding IsEnabled}">
-                        <TextBox Text="{Binding Hour, UpdateSourceTrigger=PropertyChanged}" KeyDown="KeyDownHandler">
+                        <TextBox Text="{Binding Hour, UpdateSourceTrigger=PropertyChanged}">
+                            <wi:Interaction.Triggers>
+                                <wi:EventTrigger  EventName="KeyDown">
+                                    <i:CallMethodAction
+                                        TargetObject="{Binding}"
+                                        MethodName="KeyDownHandler"/>
+                                </wi:EventTrigger>
+                            </wi:Interaction.Triggers>
                         </TextBox>
                         <StackPanel Orientation="Vertical" VerticalAlignment="Stretch">
                             <Button Command="{Binding ClickUpHour}" Content="˄" Height="32" Width="20" />
@@ -58,7 +69,15 @@
                     </StackPanel>
                     <Label Content=":" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" FontSize="24"/>
                     <StackPanel Orientation="Horizontal" IsEnabled="{Binding IsEnabled}">
-                        <TextBox Text="{Binding Minute, UpdateSourceTrigger=PropertyChanged}" KeyDown="KeyDownHandler"/>
+                        <TextBox Text="{Binding Minute, UpdateSourceTrigger=PropertyChanged}">
+                            <wi:Interaction.Triggers>
+                                <wi:EventTrigger EventName="KeyDown">
+                                    <i:CallMethodAction
+                                        TargetObject="{Binding}"
+                                        MethodName="KeyDownHandler"/>
+                                </wi:EventTrigger>
+                            </wi:Interaction.Triggers>
+                        </TextBox>
                         <StackPanel Orientation="Vertical" VerticalAlignment="Stretch">
                             <Button Command="{Binding ClickUpMinute}" Content="˄" Height="32" Width="20" />
                             <Button Command="{Binding ClickDownMinute}" Content="˅" Height="32" Width="20" />
@@ -71,7 +90,7 @@
                             <Image Source="/AlarmClock;component/resources/btns/add.png" Height="24" Style="{StaticResource ImageEnabled}"/>
                         </Button>
                         <Separator Style="{StaticResource {x:Static ToolBar.SeparatorStyleKey}}"
-                            Visibility="{Binding IsSaveEnabled, Converter={StaticResource BoolToVisibility}}"/>
+                            Visibility="{Binding IsRingEnabled, Converter={StaticResource BoolToVisibility}}"/>
                         <Button Style="{StaticResource ToolButton}" HorizontalAlignment="Center" 
                             ToolTip="Ring alarm clock" Command="{Binding RingAlarm}"
                             Visibility="{Binding IsRingEnabled, Converter={StaticResource BoolToVisibility}}">
@@ -104,26 +123,33 @@
 
         <TextBlock Grid.Row="0" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Center"
                    Text="{Binding CurrentTime, UpdateSourceTrigger=PropertyChanged}" FontSize="24"/>
-        <StackPanel Orientation="Horizontal" Grid.Row="0" Grid.Column="2">
+        <StackPanel Orientation="Horizontal" Grid.Row="0" Grid.Column="1" HorizontalAlignment="Center">
             <Image Source="/AlarmClock;component/resources/btns/man-user.png" Height="32"/>
             <Label Content="{Binding CurrentUser.Login}" HorizontalContentAlignment="Center" 
                    VerticalContentAlignment="Center" FontSize="24"/>
-
-            <Button Style="{StaticResource ToolButton}" ToolTip="Sign Out" 
-                    Margin="20,0,0,0" Command="{Binding SignOut}" Name="SignOut">
-                <Image Source="/AlarmClock;component/resources/btns/signout.png" Height="32"/>
-            </Button>
         </StackPanel>
-
-        <StackPanel Grid.Row="1" Grid.Column="1" Orientation="Vertical">
+        <Button Grid.Row="0" Grid.Column="2" Style="{StaticResource ToolButton}" ToolTip="Sign Out" 
+            Margin="20,0,0,0" Command="{Binding SignOut}" Name="SignOut" HorizontalAlignment="Center">
+            <Image Source="/AlarmClock;component/resources/btns/signout.png" Height="32"/>
+        </Button>
+        <Grid Grid.Row="1" Grid.Column="1">
             <ListBox ItemsSource="{Binding AlarmClocks}" SelectedItem="{Binding SelectedAlarm}"
-                     BorderThickness="0" HorizontalContentAlignment="Center">
+                     BorderThickness="0" HorizontalContentAlignment="Center"
+                     ScrollViewer.VerticalScrollBarVisibility="Auto">
                 <ListBox.ItemTemplate>
                     <DataTemplate>
                         <ContentControl Template="{StaticResource AlarmClockTemplate}"/>
                     </DataTemplate>
                 </ListBox.ItemTemplate>
             </ListBox>
-        </StackPanel>
+        </Grid>
+        <StatusBar Grid.Column="0" Grid.Row="2" Grid.ColumnSpan="3">
+            <StatusBarItem>
+                <TextBlock Text="{Binding AlarmClocks.Count, Converter={StaticResource CountConverter}}"/>
+            </StatusBarItem>
+        </StatusBar>
     </Grid>
 </UserControl>
+<!--add class CountConverter for count existing user's clocks
+fixed problem with displaying user name
+add basic scroll-->
